refactor : checkList 패키지 구조 분리
