+++ src/main/java/com/workhub/checklist/controller/CheckListController.java
@@ -5,9 +5,9 @@ import com.workhub.checklist.dto.CheckListCreateRequest;
 import com.workhub.checklist.dto.CheckListItemStatus;
 import com.workhub.checklist.dto.CheckListResponse;
 import com.workhub.checklist.dto.CheckListUpdateRequest;
-import com.workhub.checklist.service.CreateCheckListService;
-import com.workhub.checklist.service.ReadCheckListService;
-import com.workhub.checklist.service.UpdateCheckListService;
+import com.workhub.checklist.service.checkList.CreateCheckListService;
+import com.workhub.checklist.service.checkList.ReadCheckListService;
+import com.workhub.checklist.service.checkList.UpdateCheckListService;
 import com.workhub.global.response.ApiResponse;
 import com.workhub.global.security.CustomUserDetails;
 import jakarta.validation.Valid;
