+++ src/test/java/com/workhub/checklist/service/CreateCheckListServiceTest.java
@@ -5,6 +5,7 @@ import com.workhub.checklist.entity.CheckList;
 import com.workhub.checklist.entity.CheckListItem;
 import com.workhub.checklist.entity.CheckListOption;
 import com.workhub.checklist.entity.CheckListOptionFile;
+import com.workhub.checklist.service.checkList.CreateCheckListService;
 import com.workhub.global.error.ErrorCode;
 import com.workhub.global.error.exception.BusinessException;
 import org.junit.jupiter.api.BeforeEach;
